backend/requirements.txt
asgiref==3.11.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distlib==0.4.0
dj-database-url==3.1.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.2
fonttools==4.59.2
idna==3.10
iniconfig==2.1.0
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.3.2
packaging==25.0
pillow==11.3.0
platformdirs==4.5.1
pluggy==1.6.0
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.4.0
virtualenv==20.35.4
